Import correct slackbot_settings.py during tests
If a slackbot_settings.py is in the same dir as run.py, it
was imported during tests rather than the one in the functional
dir. This change looks for the test environment variable and updates
sys.path to ensure that the test settings file is imported first
rather than any locally defined settings file.