Update send_webapi() to send as_user by default
Slackbot has an option to define an icon and emoji for web API sends. Each
send with the web API allows the caller to specify a different icon/emoji.
This strikes me as behavior that probably shouldn't be default. By using the
as_user option, the send keeps the same icon/emoji as the bot integration and
looks identical to messages sent with .reply().