Import correct slackbot_settings.py during tests (#70)
If a slackbot_settings.py is in the same dir as run.py, it
was imported during tests rather than the one in the functional dir.
This change adds a run.py in the functional dir and sets up PYTHONPATH
so run.py is able to import the things it needs to, while also ensuring
that the correct slackbot_settings is imported.

Unit tests run in the context of py.test, so fixing them was a different
challenge. pytest's monkeypatch can be used to mock out a module, but
only in recent versions of py.test. As a result, I had to add pytest
to requirements.txt in order for tests to work on Travis.