Add timeout setting for Slacker (#175)
* Add timeout setting for Slacker

There are serveral parameters for slacker:
slacker.Slacker(token,
                incoming_webhook_url=None,
                timeout=DEFAULT_TIMEOUT,
                http_proxy=None,
                https_proxy=None,
                session=None,
                rate_limit_retries=DEFAULT_RETRIES)
However, in the previous code, only token is passed to slacker, and
timeout could not be set. It uses DEFAULT_TIMEOUT in slacker, which
is 10 seconds, not enough for slackbot to upload large file (20MB
video for example).
In this commit, timeout could be set in settings as token. Bot()
would pass it to SlackClient and to Slacker if TIMEOUT exists in
settings.

* Add test for timeout