Add support for `bot_message` events without usernames
As per https://api.slack.com/events/message/bot_message a `bot_message`
may not always contain a username. Username is only present should the
username be overrided in the event request.

Currently when SlackBot receives a `bot_message` with no overriden username it exits not passing the event forward.

This changes updated the relevant code to pull the username from the `bot_profile`